chore: rename sn-api internal transaction  file
