+++ crates/starknet_api/src/lib.rs
@@ -8,8 +8,8 @@ pub mod core;
 pub mod crypto;
 pub mod data_availability;
 pub mod deprecated_contract_class;
+pub mod executable_transaction;
 pub mod hash;
-pub mod internal_transaction;
 pub mod rpc_transaction;
 pub mod serde_utils;
 pub mod state;
